requirements/base.txt
Django==1.7.0
bpython==0.12
django-braces==1.4.0
django-model-utils==2.0.3
django-autoslug==1.7.2
django-widget-tweaks==1.3
django-admin-honeypot==0.3.0
django-ckeditor-updated==4.4.0
django-imagekit==3.2.1
django-pipeline==1.3.24
jsmin==2.0.9
logutils==0.3.3
South==0.8.2
Pillow==2.4.0
pytz==2014.4
